Add protection fuses and diodes to LVDS interface.
I'm a bit concerned about how the fuses will affect the data lines
when the device is in pass-through mode.  However, I'm also somewhat
reluctanct to switch to using a relay.

An extreme option is to use some analogue switches powered with, say,
a +/- 16V rail...  Seems like overkill though.