Add LEDs and connect CBUS lines to LEDs and MSP
The FT232 has a number of configurable "CBUS" lines.  These lines can
be configured to do a variety of things.

libftdi (currently) doesn't support configuring the CBUS pins, as FTDI
require one to enter into an NDA to discover the EEPROM mapping (some
some undisclosed, and presumably completely insane, reason).  The
FT232 comes with a factory preset CBUS pin configuration, which this
schematic uses.  The TXLED and RXLED CBUS pins are connected to an LED
each, and indicate when data is being sent and received through USB.

The MSP needs to know when the USB device has gone into suspend so
that it can power down the SMPS, and sleep itself.  (Although it
should be possible for the MSP to be in sleep all the time, as it'll
be a very interrupt-driven beast.)

In the event that we do, at some point in the future, reverse engineer
the FT232 EEPROM contents (which looks like a trivial task), we could
reconfigure the CBUS pins to do something magic for a feature that I
can't yet think of.  So, I've connected all the CBUS pins through to
*interruptable* IO lines on the MSP.

Note that there are no power LEDs.  The current budget for USB suspend
mode is 2.5mA -- too little for any LEDs really.  I've also included
two debug LEDs that are connected to the MSP.